refactor: extract shared sport page logic into useSportPage hook
Move ~400 lines of duplicated state management, data loading, filtering,
and pagination logic from app/page.tsx and app/nhl/page.tsx into a shared
useSportPage hook. Extract LiveGamesSection and GameDateGrid as reusable
components. Each page now only contains sport-specific config and rendering.
